requirements.txt
aiogram>=3.4.1
asyncpg>=0.29.0
python-dotenv>=1.0.1
redis>=5.0.1
sqlalchemy>=2.0.25
greenlet>=3.0.0
supabase>=2.4.0
sentry-sdk>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
